Add files via upload
This program allows the user to create their own 2-body system as a means to show how orbits between two objects work. If the user makes a planet orbit the primary star, it will be clear that most planetary orbits are being circularized where the orbit looks like a circle compared to very elliptical. Using parameters for planets in our solar system shows that the orbits are almost perfect circles but it can be seen that they are in fact somewhat elliptical. Using a binary star system will better show how the center of mass is what two objects orbit and how the velocities and distances change during an orbit as there are greater changes for this case - results which can be boiled down to the masses being much closer to each other. To make the simulation more appealing to the eye, there is a randomized star background as well as lines that follow the object, tracing out the orbit as they move around the center of mass.